add rmsnorm polciy
